refactor: fixed the paths
